adding interest_rate and balance to bills

+++ db/migrate/20211215142014_create_bills.rb
@@ -2,7 +2,7 @@ class CreateBills < ActiveRecord::Migration[6.1]
   def change
     create_table :bills do |t|
       t.string :title
-      t.integer :amount_due
+      t.money :amount_due
       t.integer :due_date
       t.references :user, null: false, foreign_key: true
